switched to one byte for class/method ids
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@578891 13f79535-47bb-0310-9956-ffa450edef68

+++ java/common/src/main/java/org/apache/qpidity/transport/network/Assembler.java
@@ -165,7 +165,7 @@ public class Assembler implements Receiver<NetworkEvent>, NetworkDelegate
         switch (type)
         {
         case Frame.METHOD:
-            int methodType = (int) dec.readLong();
+            int methodType = dec.readShort();
             Method method = Method.create(methodType);
             method.read(dec, major, minor);
             return method;
